- Add CalculoController
Foi adicionado a controller que chama CalculoService injetado que calcula o CDB rebendo como parâmetro um valor monetário e uma quantidade de meses.

O método Get da CalculoController retorna um objeto contendo o valor bruto e o valor liquido do CDB calculado

+++ Avaliacao-Atos.Application/Interfaces/ICalculoService.cs
@@ -7,6 +7,6 @@ namespace Avaliacao_Atos.Application.Interfaces
 {
     public interface ICalculoService
     {
-        CalculoViewModel GetCalculo(double valor_monetario, int qtd_meses);
+        CalculoViewModel GetCalculo(decimal valor_monetario, int qtd_meses);
     }
 }
